Move package_info into constants
