+++ compile_utils/constants.py
@@ -1,4 +1,11 @@
 """Constant values used during compilation, such as file names"""
 
+# Project
+
+IMAGE_VIEWER_NAME: str = "image_viewer"
+PROJECT_FILE: str = "pyproject.toml"
+
+# Compilation
+
 BUILD_INFO_FILE: str = "build_info.txt"
 REPORT_FILE: str = "compilation-report.xml"
