--- compile_utils/package_info.py
@@ -1,4 +1,0 @@
-"""Information on the image viewer package"""
-
-IMAGE_VIEWER_NAME: str = "image_viewer"
-PROJECT_FILE: str = "pyproject.toml"
